# docs/01_contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Exhibition Infrastructure API
  version: 0.1.0
  description: >
    Phase 2A OpenAPI skeleton for Flutter App canonical paths, plus retained Admin and
    internal service alignment placeholders.
servers:
  - url: http://localhost:3000
    description: Local BFF
  - url: http://localhost:3001
    description: Local Server
...
paths:
...
  /api/app/forum/feed:
    get:
      tags: [Forum]
      summary: Forum feed canonical path for Flutter App
      operationId: getAppForumFeed
      parameters:
        - in: query
          name: scope
          required: true
          schema:
            $ref: '#/components/schemas/ForumFeedScope'
        - in: query
          name: topicId
          required: false
          schema:
            type: string
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: >
            Minimum forum feed aggregation surface for Flutter App. This path
            carries ordered post-feed items for the current `square / local /
            following` browse chain, plus the frozen cursor page carrier and
            internal topic-label/filter semantics. It must not become forum
            truth, moderation truth, ranking configuration, or a second
            discovery state machine.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumFeedResponse'
...
  /api/app/forum/topic/list:
    get:
      tags: [Forum]
      summary: Forum topic list canonical path for Flutter App
      operationId: getAppForumTopicList
      parameters:
        - in: query
          name: categoryKey
          required: false
          schema:
            type: string
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: >
            Minimum forum topic list aggregation for Flutter App. This path
            keeps the same frozen topic-card semantics as `forum/feed` and adds
            only category filtering plus cursor pagination.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumTopicListResponse'
...
  /api/app/forum/publish:
    post:
      tags: [Forum]
      summary: Publish existing forum draft via canonical path for Flutter App
      operationId: publishAppForumDraft
      requestBody:
        required: true
        description: >
          Minimum forum publish handoff for Flutter App. This path forwards only
          an existing Server-owned `draftId`; it must not create a second draft
          lifecycle, edit truth, or attachment upload path family.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForumPublishRequest'
      responses:
        "202":
          description: >
            Forum publish command accepted as a minimum handoff only. The server
            returns the minimum published topic/post anchors and summary needed
            by Flutter App continuation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumPublishAcceptedResponse'
        "400":
          description: >
            Controlled forum publish request error. If `draftId` is missing or
            the request body violates the minimum publish boundary, the server
            must return `FORUM_PUBLISH_INVALID`.
        "409":
          description: >
            Controlled forum publish unavailable response. If the current draft
            state or upstream topic state does not allow publish, the server
            must return `FORUM_PUBLISH_INVALID_STATE`.
  /api/app/forum/draft/save:
    post:
      tags: [Forum]
      summary: Save forum post draft via canonical path for Flutter App
      operationId: saveAppForumDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForumDraftSaveRequest'
      responses:
        "202":
          description: Minimum forum draft-save accepted response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumDraftSavedResponse'
        "400":
          description: >
            Controlled draft-save request error. Invalid current draft input
            must return `FORUM_DRAFT_INVALID`.
  /api/app/forum/draft/list:
    get:
      tags: [Forum]
      summary: Forum draft list canonical path for Flutter App
      operationId: getAppForumDraftList
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: >
            Minimum forum draft list aggregation for Flutter App. This path
            shapes only the actor-visible draft cards and frozen cursor page
            carrier; it must not expose edit history, autosave timeline, or
            moderation internals.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumDraftListResponse'
...
  /api/app/file/upload/init:
    post:
      tags: [Upload]
      summary: Initiate three-step upload via app-facing canonical path
      operationId: initiateAppUpload
      requestBody:
        required: true
        description: >
          The upload init request must contain the frozen fields from
          upload_contracts.yaml only: businessType, businessId, fileKind, mimeType,
          size, checksum.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInitRequest'
      responses:
        "200":
          description: >
            Upload strategy issued for step 1 of the three-step upload flow. Signed
            strategy structure remains a skeleton until transport fields are frozen.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadInitResponse'
  /api/app/file/upload/confirm:
    post:
      tags: [Upload]
      summary: Confirm uploaded object and create file truth via app-facing canonical path
      operationId: confirmAppUpload
      requestBody:
        required: true
        description: >
          Confirm request for step 3 of the upload flow. Only uploadSessionId is
          frozen for Phase 2A Flutter App consumption.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadConfirmRequest'
      responses:
        "200":
          description: >
            FileAsset confirmation skeleton. objectKey is storage location only and does
            not become business truth by itself.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadConfirmResponse'
...
  /server/uploads/confirm:
    post:
      tags: [Upload]
      summary: Internal truth upload-confirm pair for app-facing canonical upload confirm
      operationId: confirmUpload
      requestBody:
        required: true
        description: >
          Internal truth pair for `POST /api/app/file/upload/confirm`. This
          step confirms shared upload truth and freezes confirmed `FileAsset`
          output; `objectKey` remains storage location only.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadConfirmRequest'
      responses:
        "200":
          description: >
            Internal FileAsset confirmation response. Confirm is the only step
            in the shared upload corridor that creates the `FileAsset` business
            truth reference.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadConfirmResponse'
...
components:
  schemas:
...
    ForumCursorPageInfo:
      type: object
      required:
        - nextCursor
        - hasMore
      description: >
        Frozen cursor page carrier for forum list-like paths only. This is a
        transport paging projection and not forum ranking, indexing, or
        moderation truth.
      properties:
        nextCursor:
          type:
            - string
            - 'null'
        hasMore:
          type: boolean
    ForumFeedScope:
      type: string
      enum:
        - square
        - local
        - following
      description: >
        Current frozen first-level forum feed scopes only.
    ForumAuthorSummary:
      type: object
      required:
        - authorId
        - displayName
      description: >
        Minimum forum author summary only. This object is not profile truth,
        organization management truth, or a second identity contract.
      properties:
        authorId:
          type: string
        displayName:
          type: string
        organizationName:
          type:
            - string
            - 'null'
    ForumAttachmentRef:
      type: object
      required:
        - fileAssetId
        - fileName
        - mimeType
      description: >
        Confirmed file truth reference only for forum read models. `objectKey`
        must stay outside the app-facing forum business contract.
      properties:
        fileAssetId:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
    ForumEngagementSummary:
      type: object
      required:
        - replyCount
        - likeCount
        - viewCount
      description: >
        Minimum forum engagement projection only. This is not ranking truth or
        operator analytics truth.
      properties:
        replyCount:
          type: integer
          minimum: 0
        likeCount:
          type: integer
          minimum: 0
        viewCount:
          type: integer
          minimum: 0
    ForumTopicState:
      type: string
      enum:
        - published
        - locked
        - archived
      description: >
        Current minimum topic state set consumed by the Flutter App forum
        surfaces.
    ForumPostState:
      type: string
      enum:
        - published
        - hidden
      description: >
        Current minimum post state set consumed by the Flutter App forum
        surfaces.
    ForumDraftState:
      type: string
      enum:
        - draft
        - ready_to_publish
      description: >
        Current minimum draft state set consumed by the Flutter App forum
        draft surfaces.
    ForumDraftType:
      type: string
      enum:
        - topic
        - reply
      description: >
        Current minimum draft-kind set consumed by the Flutter App forum draft
        surfaces.
...
    ForumTopicCard:
      type: object
      required:
        - topicId
        - title
        - excerpt
        - categoryKey
        - state
        - author
        - engagement
        - lastActiveAt
      description: >
        Shared minimum topic-card projection for `forum/feed`,
        `forum/topic/list`, and the recent-topic carrier inside `forum/me/index`.
      properties:
        topicId:
          type: string
        title:
          type: string
        excerpt:
          type: string
        categoryKey:
          type: string
        state:
          $ref: '#/components/schemas/ForumTopicState'
        author:
          $ref: '#/components/schemas/ForumAuthorSummary'
        engagement:
          $ref: '#/components/schemas/ForumEngagementSummary'
        lastActiveAt:
          type: string
          format: date-time
        highlightedPostId:
          type:
            - string
            - 'null'
...
    ForumFeedItem:
      type: object
      required:
        - postId
        - topicId
        - topicLabel
        - title
        - excerpt
        - state
        - author
        - engagement
        - publishedAt
        - viewerHasLiked
        - viewerHasBookmarked
        - viewerFollowsTopic
      description: >
        Minimum post-centric feed item for the current forum browse chain.
      properties:
        postId:
          type: string
        topicId:
          type: string
        topicLabel:
          type: string
        title:
          type: string
        excerpt:
          type: string
        state:
          $ref: '#/components/schemas/ForumPostState'
        author:
          $ref: '#/components/schemas/ForumAuthorSummary'
        engagement:
          $ref: '#/components/schemas/ForumEngagementSummary'
        publishedAt:
          type: string
          format: date-time
        viewerHasLiked:
          type: boolean
        viewerHasBookmarked:
          type: boolean
        viewerFollowsTopic:
          type: boolean
...
    ForumDraftCard:
      type: object
      required:
        - draftId
        - draftType
        - title
        - excerpt
        - state
        - updatedAt
        - attachmentRefs
      description: >
        Minimum forum draft-card projection for `forum/draft/list` and the
        me-subtree recent-draft carrier only.
      properties:
        draftId:
          type: string
        draftType:
          $ref: '#/components/schemas/ForumDraftType'
        topicId:
          type:
            - string
            - 'null'
        title:
          type: string
        excerpt:
          type: string
        state:
          $ref: '#/components/schemas/ForumDraftState'
        updatedAt:
          type: string
          format: date-time
        attachmentRefs:
          type: array
          items:
            $ref: '#/components/schemas/ForumAttachmentRef'
...
    ForumPublishRequest:
      type: object
      required:
        - draftId
      description: >
        Minimum forum publish handoff request only. Upload signing remains on
        the shared file-upload path family; forum publish forwards only an
        existing Server-owned `draftId`.
      properties:
        draftId:
          type: string
    ForumPublishSummary:
      type: object
      required:
        - title
        - publishedAt
      description: >
        Controlled forum publish summary skeleton only.
      properties:
        title:
          type: string
        publishedAt:
          type: string
          format: date-time
    ForumPublishAcceptedResponse:
      type: object
      required:
        - draftId
        - topicId
        - postId
        - state
        - summary
      description: >
        Minimum forum publish command success body only. This response carries
        the published anchors required by the current Flutter App continuation
        and must not expand into forum detail truth.
      properties:
        draftId:
          type: string
        topicId:
          type: string
        postId:
          type: string
        state:
          type: string
          enum:
            - published
        summary:
          $ref: '#/components/schemas/ForumPublishSummary'
...
    ForumDraftSaveRequest:
      type: object
      required:
        - topicId
        - title
        - body
      description: >
        Current minimum post-draft save request only. Attachment references
        must use confirmed `FileAsset` ids only.
      properties:
        draftId:
          type:
            - string
            - 'null'
        topicId:
          type: string
        title:
          type: string
        body:
          type: string
        attachmentFileAssetIds:
          type: array
          items:
            type: string
    ForumDraftSavedResponse:
      type: object
      required:
        - draftId
        - state
        - updatedAt
      properties:
        draftId:
          type: string
        state:
          $ref: '#/components/schemas/ForumDraftState'
        updatedAt:
          type: string
          format: date-time
...
    ForumFeedResponse:
      type: object
      required:
        - items
        - page
      description: >
        Minimum forum feed carrier for Flutter App.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ForumFeedItem'
        page:
          $ref: '#/components/schemas/ForumCursorPageInfo'
...
    ForumTopicListResponse:
      type: object
      required:
        - items
        - page
      description: >
        Minimum forum topic-list carrier for Flutter App.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ForumTopicCard'
        page:
          $ref: '#/components/schemas/ForumCursorPageInfo'
    ForumDraftListResponse:
      type: object
      required:
        - items
        - page
      description: >
        Minimum forum draft-list carrier for Flutter App.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ForumDraftCard'
        page:
          $ref: '#/components/schemas/ForumCursorPageInfo'
...
    UploadInitRequest:
      type: object
      required:
        - businessType
        - businessId
        - fileKind
        - mimeType
        - size
        - checksum
      description: >
        Minimum shared upload init request for the current three-step upload
        corridor only. For the project publish minimum corridor, SSOT freezes
        the bound publish upload combination to `businessType=project` and
        `fileKind=evidence`; `businessId` remains a carried binding field and
        may be `null` before project truth is accepted.
      properties:
        businessType:
          type: string
        businessId:
          type:
            - string
            - 'null'
        fileKind:
          type: string
        mimeType:
          type: string
        size:
          type: integer
          minimum: 1
        checksum:
          type: string
    UploadDirectUploadStrategy:
      type: object
      required:
        - url
        - method
      description: >
        Direct upload transport directive only. This structure carries byte
        transport instructions and must not be interpreted as business truth.
      properties:
        url:
          type: string
        method:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
    UploadConfirmEndpoint:
      type: object
      required:
        - endpoint
      description: >
        Confirm handoff descriptor only for step 3 of the shared upload
        corridor. This must not become a second upload family.
      properties:
        endpoint:
          type: string
    UploadInitResponse:
      type: object
      required:
        - uploadSessionId
        - directUpload
        - confirm
      description: >
        Minimum upload-init strategy response for the shared
        `init -> direct upload -> confirm` corridor. No business truth is
        created until confirm succeeds.
      properties:
        uploadSessionId:
          type: string
        directUpload:
          $ref: '#/components/schemas/UploadDirectUploadStrategy'
        confirm:
          $ref: '#/components/schemas/UploadConfirmEndpoint'
    UploadConfirmRequest:
      type: object
      required:
        - uploadSessionId
      description: >
        Minimum shared upload confirm request only. This request identifies the
        previously issued upload session for FileAsset truth confirmation.
      properties:
        uploadSessionId:
          type: string
    UploadConfirmResponse:
      type: object
      required:
        - fileAssetId
      description: >
        Minimum FileAsset confirmation response only. This response freezes the
        confirmed `FileAsset` truth reference by id; `objectKey` remains
        storage location only and stays outside the business contract.
      properties:
        fileAssetId:
          type: string
...